+++ Includes/minishell.h
@@ -243,6 +243,7 @@ void		free_dlist(t_dlist **head);
 void		free_struct(t_mshell *shell);
 void		free_arr(char **arr);
 void		free_export(t_envp **export);
+void		free_errors(t_error *error);
 void		multifree(void *ptr1, void *ptr2, void *ptr3, void *ptr4);
 
 //UTILS
