+++ Sources/Execution/execution.c
@@ -14,55 +14,6 @@
 
 extern long long	g_status;
 
-int	is_builtin(t_tokens *temp)
-{
-	if (!ft_strcmp(temp->cmd_arr[0], "echo")
-		|| !ft_strcmp(temp->cmd_arr[0], "cd")
-		|| !ft_strcmp(temp->cmd_arr[0], "exit")
-		|| !ft_strcmp(temp->cmd_arr[0], "env")
-		|| !ft_strcmp(temp->cmd_arr[0], "unset")
-		|| !ft_strcmp(temp->cmd_arr[0], "pwd")
-		|| !ft_strcmp(temp->cmd_arr[0], "export"))
-		return (g_status = 0, 1);
-	return (0);
-}
-
-void	builtin_forwarding_pipe(t_tokens *temp, t_mshell *shell)
-{
-	if (!ft_strcmp(temp->cmd_arr[0], "echo"))
-		echo_case(temp->cmd_arr, 1);
-	else if (!ft_strcmp(temp->cmd_arr[0], "cd"))
-		cd_case(shell, temp->cmd_arr);
-	else if (!ft_strcmp(temp->cmd_arr[0], "exit"))
-		exit_case(shell, temp->cmd_arr);
-	else if (!ft_strcmp(temp->cmd_arr[0], "env"))
-		env_case(shell, temp->cmd_arr, 1);
-	else if (!ft_strcmp(temp->cmd_arr[0], "unset"))
-		unset_case(shell, temp->cmd_arr);
-	else if (!ft_strcmp(temp->cmd_arr[0], "pwd"))
-		pwd_case(shell, temp->cmd_arr, 1);
-	else if (!ft_strcmp(temp->cmd_arr[0], "export"))
-		export_case(shell, temp->cmd_arr, 1);
-}
-
-void	builtin_forwarding(t_tokens *temp, t_mshell *shell)
-{
-	if (!ft_strcmp(temp->cmd_arr[0], "echo"))
-		echo_case(temp->cmd_arr, temp->fd_out);
-	else if (!ft_strcmp(temp->cmd_arr[0], "cd"))
-		cd_case(shell, temp->cmd_arr);
-	else if (!ft_strcmp(temp->cmd_arr[0], "exit"))
-		exit_case(shell, temp->cmd_arr);
-	else if (!ft_strcmp(temp->cmd_arr[0], "env"))
-		env_case(shell, temp->cmd_arr, temp->fd_out);
-	else if (!ft_strcmp(temp->cmd_arr[0], "unset"))
-		unset_case(shell, temp->cmd_arr);
-	else if (!ft_strcmp(temp->cmd_arr[0], "pwd"))
-		pwd_case(shell, temp->cmd_arr, temp->fd_out);
-	else if (!ft_strcmp(temp->cmd_arr[0], "export"))
-		export_case(shell, temp->cmd_arr, temp->fd_out);
-}
-
 void	executable(t_tokens *temp, t_mshell *shell)
 {
 	pid_t	child;
@@ -87,15 +38,13 @@ void	executable(t_tokens *temp, t_mshell *shell)
 		if (waitpid(child, (int *)&g_status, 0) == -1)
 			return (perror("waitpid"), free_struct(shell), exit(EXIT_FAILURE));
 	}
-	get_fork_status();
 	if (temp->cmd_arr)
 		free_arr(temp->cmd_arr);
-	temp->cmd_arr = NULL;
+	return (temp->cmd_arr = NULL, get_fork_status());
 }
 
 void	exec_forwarding(t_tokens *temp, t_mshell *shell)
 {
-	// ft_dprintf(2, "in exec forwarding, cmd = %s, fd = %d\n", temp->cmd_arr[0], temp->fd_out);
 	if (has_bad_fd(temp))
 	{
 		g_status = 1;
