+++ Sources/Free/multifree.c
@@ -6,7 +6,7 @@
 /*   By: chonorat <chonorat@student.42lyon.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/22 14:02:20 by chonorat          #+#    #+#             */
-/*   Updated: 2023/10/22 14:24:04 by chonorat         ###   ########.fr       */
+/*   Updated: 2023/10/23 15:10:32 by chonorat         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,4 +22,19 @@ void	multifree(void *ptr1, void *ptr2, void *ptr3, void *ptr4)
 		free(ptr3);
 	if (ptr4)
 		free(ptr4);
+}
+
+void	free_errors(t_error *error)
+{
+	t_error	*temp;
+
+	while (error)
+	{
+		if (error->content)
+			free(error->content);
+		temp = error;
+		error = temp->next;
+		free(temp);
+	}
+	error = NULL;
 }
